requirements.txt
Adafruit-GPIO==1.0.3
Adafruit-MCP3008==1.0.2
Adafruit-PureIO==0.2.1
Adafruit-SSD1306==1.6.1
appdirs==1.4.3
cookies==2.2.1
cycler==0.10.0
decorator==4.0.11
feedparser==5.1.3
forecastiopy==0.1
html5lib==0.999
httplib2==0.9.1
matplotlib==2.0.0
networkx==1.11
numpy==1.12.1
packaging==16.8
pbr==2.0.0
pkg-resources==0.0.0
pyparsing==2.2.0
PyQt5==5.8.1
pyserial==3.3
python-dateutil==2.6.0
python-forecastio==1.3.5
pytz==2016.10
requests==2.9.1
responses==0.5.1
scikit-fuzzy==0.2
scipy==0.19.0
sip==4.19.2
six==1.10.0
spidev==3.2
stevedore==1.21.0
ubidots==1.6.6
urllib3==1.13.1
virtualenv==15.0.1
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
